python-src/pyproject.toml
[project]
name = "python-src"
version = "0.1.0"
description = "FastAPI server for transformer pipelines"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "transformers>=4.35.2",
    "torch>=2.1.0",
    "pillow>=10.1.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0"
]
